add tests to repository

+++ product-api/src/main/java/com/qima/product/product/api/dto/ProductPatchRequest.java
@@ -1,5 +1,6 @@
 package com.qima.product.product.api.dto;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public record ProductPatchRequest(
@@ -7,6 +8,6 @@ public record ProductPatchRequest(
         Optional<String> description,
         Optional<Long> categoryId,
         Optional<Boolean> available,
-        Optional<Double> price) {
+        Optional<BigDecimal> price) {
 
 }
